\name{design}
\alias{design}
\docType{data}
\title{ Experimental design }
\description{ A vector of integers specifying for each column(sample) to which group number it belongs.}
\usage{data(design)}
\format{
  The format is:
 num [1:16] 0 0 0 0 0 0 0 0 1 1 1 1 1 1 1 1
}
\details{ In the example above there are two groups. Each group contains 8 samples}
\examples{
data(design)
}
\keyword{datasets}
