\name{fdr-internal}
\alias{compute.f.statistic}
\alias{compute.statistic}
\alias{compute.t.statistic}
\alias{fdr.adaptive.c}
\alias{fdr.adaptive.c.resampling}
\alias{fdr.basic.comp}
\alias{fdr.bh}
\alias{fdr.error}
\alias{compute.resampling.stat}
\alias{compute.statistic}
\alias{compute.t.statistic}
\alias{fdr.plot}
\alias{fdr.pt}
\alias{fdr.q}
\alias{fdr.qu} 
\alias{gene.expression.normalization} 
\alias{get.resampling.statistic.array}     
\alias{OnBrowseDesign} 
\alias{OnBrowseInput} 
\alias{OnBrowseOutput} 
\alias{OnCompute} 
\alias{OnRbfa1} 
\alias{OnRbfa2} 
\alias{OnRbfa3}
\alias{OnRbfa4}
\alias{OnRbpm1}
\alias{OnRbpm2}
\alias{OnCancel}
\alias{OnOK}
\alias{OnHelp}
\alias{read.matrix}   


\title{Internal Functions in the Fdr Package}
\description{
  Internal functions in the fdr package.
}

\keyword{internal}
