\name{fdr.gui}
\alias{fdr.gui}
\title{ fdr.gui}
\description{This function runs the fdrame Graphic User Interface. }
\usage{fdr.gui()
}
\references{Reiner A, Yekutieli D, Benjamini Y: Identifying differentially expressed genes using false discovery rate controlling procedures. Bioinformatics 19:368-375, 2003
Benjamini, Y., Krieger, A.M.,Yekutieli, D. (2001) "Two Staged Linear Step Up FDR Controlling Procedure", Technical Report Department of Statistics and O.R., Tel Aviv University.
}
\author{Yoav Benjamini, Effi Kenigsberg, Anat Reiner, Daniel Yekutieli}
\examples{
fdr.gui()
}
\keyword{htest}